// pro/src/apiClient/adage/models/ListCollectiveOfferTemplateResponseModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CollectiveOfferTemplateResponseModel } from './CollectiveOfferTemplateResponseModel';
export type ListCollectiveOfferTemplateResponseModel = {
  collectiveOffers: Array<CollectiveOfferTemplateResponseModel>;
};